ref step 5.1 fix signin - another browser

+++ models/auth.go
@@ -49,7 +49,7 @@ func (u User) Signup(w http.ResponseWriter, r *http.Request) {
 }
 
 //Signin function dsds
-func (uStr *User) Signin(w http.ResponseWriter, r *http.Request) {
+func (uStr *User) Signin(w http.ResponseWriter, r *http.Request, s general.Session) {
 
 	var isUserOrEmail bool
 
@@ -67,46 +67,47 @@ func (uStr *User) Signin(w http.ResponseWriter, r *http.Request) {
 	//if cookie timeout -> delete user from system
 
 	if utils.AuthType == "default" {
+
 		if !isUserOrEmail {
-			log.Println("1")
 			err = DB.QueryRow("SELECT id, password FROM users WHERE email=?", uStr.Email).Scan(&user.ID, &user.Password)
 			if err != nil {
 				log.Println("err email")
 				utils.AuthError(w, r, err, "user by Email not found", utils.AuthType)
 				return
 			}
-			log.Println("2")
 			//email || username _> have session ? -> delete session and cookie, ressession Call
-			var uid int
-
-			err = DB.QueryRow("SELECT id FROM  users WHERE email=?", uStr.Email).Scan(&uid)
-
-			if err != nil {
-				log.Println(err, "no have user by email")
-				return
-			}
 
 			var sid int
-			err = DB.QueryRow("SELECT id FROM session WHERE user_id=?", uid).Scan(&sid)
+			err = DB.QueryRow("SELECT id FROM session WHERE user_id=?", user.ID).Scan(&sid)
 			if err != nil {
 				log.Println(err, "no have session by uid")
 				//return
 			} else {
-				log.Println("mean -> user in system, Delete cookie")
-				//get by email -> session, if have session -> drop session -> ReLogin
-				// if same browser delete cookie
-				utils.DeleteCookie(w)
-				_, err = DB.Exec("DELETE FROM session WHERE id = ?", sid)
-				//http.Redirect(w, r, "/signin", 302)
-				log.Println(sid, "sid")
-				//deleted session ->
-				if utils.AuthType == "google" {
-					_, err = http.Get("https://accounts.google.com/o/oauth2/revoke?token=" + utils.Token)
-					if err != nil {
-						log.Println(err)
+				var sessionCookie string
+				err = DB.QueryRow("SELECT uuid FROM session WHERE user_id = ?", user.ID).Scan(&sessionCookie)
+				//current Browser(previus) ==  current new - session
+				//				log.Println(s.UUID, "|||", sessionCookie, "tuki", s.UserID, user.ID)
+
+				if s.UUID == sessionCookie {
+
+					log.Println("session have, logout user,  mean -> user in system, Delete cookie")
+					//get by email -> session, if have session -> drop session -> ReLogin
+					// if same browser delete cookie
+					_, err = DB.Exec("DELETE FROM session WHERE id = ?", sid)
+					//http.Redirect(w, r, "/signin", 302)
+					//log.Println(sid, "sid")
+					//deleted session ->
+					if utils.AuthType == "google" {
+						_, err = http.Get("https://accounts.google.com/o/oauth2/revoke?token=" + utils.Token)
+						if err != nil {
+							log.Println(err)
+						}
 					}
 				}
 			}
+			//a7d1fb2e-0383-41b4-b295-25dec2bd7c0e incognito
+			// 8b4c05e6-3e03-40bf-a659-cda1dc810328,
+
 			//Logout(w, r, s)
 			//ReLogin()
 		} else if isUserOrEmail {
@@ -142,6 +143,8 @@ func (uStr *User) Signin(w http.ResponseWriter, r *http.Request) {
 	_, err = userPrepare.Exec(uuid, newSession.UserID)
 	defer userPrepare.Close()
 
+	fmt.Println(uuid, "set uuid in Db")
+
 	if err != nil {
 		utils.AuthError(w, r, err, "the user is already in the system", utils.AuthType)
 		//get ssesion id, by local struct uuid
@@ -155,6 +158,8 @@ func (uStr *User) Signin(w http.ResponseWriter, r *http.Request) {
 		log.Println(err, "her")
 		return
 	}
+
+	//update -> Middleware(Profile call) -> delete session Db, if no session Db -> exit
 
 	//if cookie not, dele
 	// this should give you time in location
